refactor: rename core/utils => core/matchers
