+++ validator/core/__init__.py
@@ -1,7 +1,6 @@
 from tokenize import Single
 from .solver import GraduationRequirementsSolver
-from .utils import SingleAssignment
-from .parser import Parser
-from . import input
+from .matchers import SingleAssignment
+from . import input, parser
 
 __all__ = ["GraduationRequirementsSolver", "SingleAssignment"]
